refactor: minor type changes
