+++ app/(pages)/admin/transactions/page.tsx
@@ -1,9 +1,8 @@
-import { TransactionDetails } from "@pft-types/index";
 import Transactions from "../components/Transactions";
 import getTransactions from "@lib/admin/getTransactions";
 
 export default async function TransactionsPage() {
-  const transactions: TransactionDetails[] = await getTransactions();
+  const transactions = await getTransactions();
 
   return <Transactions transactions={transactions} />;
 }
